feat: rename ValueVec -> Values; RefValueVec -> RefValues; + macros
yes, this is a big and ugly commits :(. I got a little carried away with
macros and this was too much of a pain to split up into bit size
commits. The general gist is ValueVec is renamed to Values and
RefValueVec is renamed to RefValues. Several macros are added to reduce
repetition and implement the same functions / traits that the previous
implementations did.